+++ docs/changes.rst
@@ -9,6 +9,7 @@ Released on ? 2015.
 * Fixed automatic translation for project owners.
 * Improved performance of commit and push operations.
 * New management command to add suggestions from command line.
+* Added support for merging comments on file upload.
 
 weblate 2.4
 -----------